refactor: split archiver/filesystem.rs into smaller files
